# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DEC",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001/api/",
            "description": "DEV"
        }
    ],
    "paths": {
...
        "/groups/{groupId}/notes": {
            "get": {
                "tags": [
                    "notes"
                ],
                "summary": "Listar todos as notas",
                "description": "Lista todas as notas: ADMINISTRADOR, REPRESENTANTE E GERENTE têm acesso a este endpoint.",
                "operationId": "8d8a903461aefae80ed93e512e523094",
                "parameters": [
                    {
                        "name": "groupId",
                        "in": "path",
                        "description": "O ID do grupo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "500": {
                        "description": "Erro"
                    },
                    "403": {
                        "description": "Não autorizado"
                    }
                }
            },
            "post": {
                "tags": [
                    "notes"
                ],
                "summary": "Criar nova nota",
                "description": "Cria uma nova nota, somente o ADMINISTRADOR e o REPRESENTANTE tem acesso a este endpoint.",
                "operationId": "75a36357e42478508d1e1d9991080c3a",
                "parameters": [
                    {
                        "name": "groupId",
                        "in": "path",
                        "description": "Id do grupo que o documento sera associado",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "titulo da nota",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "descrição da nota",
                                        "type": "string"
                                    },
                                    "color": {
                                        "description": "cor da nota desejada",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Criado"
                    },
                    "500": {
                        "description": "Erro"
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "403": {
                        "description": "Não autorizado"
                    }
                }
            }
        },
        "/groups/{groupId}/notes/{id}": {
            "get": {
                "tags": [
                    "notes"
                ],
                "summary": "Lista o registro de notas por ID",
                "description": "Lista o registro de notas por ID de referência",
                "operationId": "6cafa664ab7b94ce6e74fa6d308f8bb4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id da nota",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Note not found"
                    },
                    "200": {
                        "description": "Ok"
                    },
                    "403": {
                        "description": "Não autorizado"
                    }
                }
            },
            "put": {
                "tags": [
                    "notes"
                ],
                "summary": "Atualiza notas",
                "description": "Atualizar notas: somente o ADMINISTRADOR e o REPRESENTANTE tem acesso a este endpoint.",
                "operationId": "414de3a9c50bcc6e4def341e8798b7de",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id da nota",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "titulo da nota",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "descrição da nota",
                                        "type": "string"
                                    },
                                    "color": {
                                        "description": "cor da nota desejada",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "500": {
                        "description": "Erro"
                    },
                    "403": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Note not found"
                    }
                }
            }
        },
        "/groups/{groupId}/notes/{noteId}": {
            "delete": {
                "tags": [
                    "notes"
                ],
                "summary": "Deletar notas",
                "description": "Deletar nota por ID de referência, somente o ADMINISTRADOR e o REPRESENTANTE tem acesso a este endpoint.",
                "operationId": "02fb81a4c0f9099e7c1be92808555fb6",
                "parameters": [
                    {
                        "name": "groupId",
                        "in": "path",
                        "description": "Id do grupo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "noteId",
                        "in": "path",
                        "description": "Id da nota",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Erro"
                    },
                    "403": {
                        "description": "Não autorizado"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Authorization": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "Authorization": []
        }
    ],
...
}